+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All changes to the **Hearts of Modding** extension will be documented in this file.
 
+## [v0.25.1] - 2026-08-05
+
+### ✨ Added
+
+- **Go to definition (F12) for states.**
+
 ## [v0.25.0] - 2026-08-05
 
 ### ✨ Added